Update dependencies
bitbucket has a change in their calls:

GetContent_11(v.projectKey, runevent.Repository, path, localVarOptionals)
becomes:
GetRawContent(v.projectKey, runevent.Repository, path, localVarOptionals)

bitbucket has a change as well on how to pass files to create..

gitlab changed some of their struct to a proper type and deprecated
GetMergeRequestChanges to ListMergeRequestDiff (filled as TODO in #1470)

+++ pkg/provider/bitbucketserver/acl.go
@@ -44,7 +44,7 @@ func (v *Provider) checkOkToTestCommentFromApprovedMember(event *info.Event) (bo
 		if nextPage > 0 {
 			localVarOptionals["start"] = int(nextPage)
 		}
-		return v.Client.DefaultApi.GetActivities(v.projectKey, event.Repository, int64(v.pullRequestNumber), localVarOptionals)
+		return v.Client.DefaultApi.GetActivities(v.projectKey, event.Repository, v.pullRequestNumber, localVarOptionals)
 	})
 	if err != nil {
 		return false, err
